Move Gogs API handling to custom Ansible module
Writing an idempotent playbook which makes extensive use of the uri module is
tricky and results in some difficult to read, non-declarative constructs.
Therefore, add a gogs_users module (analogous to the existing gitlab_users
module) which encapsulates the change handling logic. The module covers a large
part of the Gogs user handling API, not all of which is currently used by our
playbook, as it is intended to be fairly generic.