Install content (online copy and slides) on server
Change server setup so that the main page is the online copy of the workshop
content created with mkdocs, and Gogs is reachable at http://servername/gogs.
This fits more naturally into the sequence of the workshop: We treat bare Git
usage first, and only later go over to pull-request workflows and portals like
Gogs.

In order to make the setup easily customizable (possibly even during a
workshop), a push-to-deploy setup is used:

1. setup.yml creates a repository FitForGit on the Gogs server
2. setup.yml creates an update hook for this repository which invokes mkdocs
   (which also gets installed, along with its dependencies)
3. setup.yml pushes the FitForGit content from the control host to the server,
   triggering the update hook

Users can then at any time make changes to the docs (either using the same copy
they installed the server from, or getting a fresh clone from the Gogs server)
and push them to the Gogs server, which will automatically update online copy
and slides.