Add page edit links
There are currently two limitations:

Using the edit link without being logged in redirects to the login page, but
Gogs does not return to the edit page after login. Thus, one has to navigate to
the correct Markdown source file, or click on the "Edit on Gogs" link again.

In contrast to pushing changes to the repository, commits made via the Gogs
interface do not trigger a rebuild of the site. Git commit hooks are not
triggered in this case either, so the only option seems to be adding a service
that gets triggered by a web hook.