Add some styling to slides and mkdocs
Markdown support for setting attributes of HTML elements (such as CSS class) is
limited to specific constructs (such as fenced code blocks), so we need to use
raw HTML tags for other elements that should get custom styling. Pandocs parses
Markdown syntax inside of HTML elements, but MkDocs does not, so we need to
fully revert to HTML here.

The light bulb icon is taken from http://simpleicon.com/dev/light-bulb-11.html,
which gives the following license:

```
By copying or downloading any icons from simpleicon, you automatically agree to the following terms and conditions:

    All trademarks of social icons, brand logos are copyright of their respective trademark owners.

You are Allowed to ...

    Use and edit these icons for any personal or commercial work without credit, however the linkback or credit for us would be very appreciated.

You are not allowed to ...

    REDISTRIBUTE OR SELL these icons on any other website or source.
```